Change rounding rules for multiple items
Previously the price for a line was determined by multiplying the price
per barrel with the number of barrels, then rounding to the nearest
penny.

With items that aren't a convenient fraction of a barrel, this could
lead to the price for two units being less than the price for two single
units.  (Eg. 1 30l keg Moravka was £58.66, but two were £117.31)

This commit changes the rounding rules such that a single item is always
priced at a whole number of pennies, and multiple items are priced at an
exact multiple of the price for the single item, so in the above case
two 30l kegs of Moravka would now be £117.32.