Save date/ref in session, add suppress_due_date flag
The date and reference for an invoice or bill is saved in the session
and restored whenever the invoice page is loaded with at least one line
item.

A "suppress_due_date" flag is added to the Contact model and, if set,
prevents the calculated due date from being sent to Xero; this is useful
for contacts where the default due date of "31 days after the
invoice/bill date" is not appropriate.

Closes #2 and #3 on github.